refactor: regroup some utils
