+++ packages/apps/reputation-oracle/server/src/modules/user/user.controller.ts
@@ -30,7 +30,7 @@ import {
 import { JwtAuthGuard } from '../../common/guards';
 import { HCaptchaGuard } from '../../common/guards/hcaptcha';
 import { RequestWithUser } from '../../common/types';
-import { prepareSignatureBody } from '../../common/utils/signature';
+import { prepareSignatureBody } from '../../utils/web3';
 import { UserService } from './user.service';
 import { Public } from '../../common/decorators';
 import { KycSignedAddressDto } from '../kyc/kyc.dto';
