+++ packages/apps/reputation-oracle/server/src/modules/webhook/webhook-outgoing.service.spec.ts
@@ -21,7 +21,7 @@ import { WebhookOutgoingService } from './webhook-outgoing.service';
 import { WebhookOutgoingEntity } from './webhook-outgoing.entity';
 import { of } from 'rxjs';
 import { HEADER_SIGNATURE_KEY } from '../../common/constants';
-import { signMessage } from '../../common/utils/signature';
+import { signMessage } from '../../utils/web3';
 import { HttpStatus } from '@nestjs/common';
 import { Web3ConfigService } from '../../common/config/web3-config.service';
 import { ServerConfigService } from '../../common/config/server-config.service';
